+++ src/main/java/net/geforcemods/securitycraft/blocks/LaserBlock.java
@@ -31,8 +31,6 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class LaserBlock extends DisguisableBlock {
 	public static final PropertyBool POWERED = PropertyBool.create("powered");
@@ -205,7 +203,6 @@ public class LaserBlock extends DisguisableBlock {
 		}
 	}
 
-	@SideOnly(Side.CLIENT)
 	@Override
 	public void randomDisplayTick(IBlockState state, World world, BlockPos pos, Random rand) {
 		if (state.getValue(POWERED)) {