+++ src/main/java/net/geforcemods/securitycraft/network/IProxy.java
@@ -1,5 +1,6 @@
 package net.geforcemods.securitycraft.network;
 
+import net.minecraft.client.particle.IParticleFactory;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
@@ -20,4 +21,6 @@ public interface IProxy {
 	}
 
 	public default void updateBlockColorAroundPosition(BlockPos pos) {}
+
+	public default void addEffect(IParticleFactory factory, World level, double x, double y, double z) {}
 }