+++ src/main/java/net/geforcemods/securitycraft/screen/CustomizeBlockScreen.java
@@ -12,6 +12,7 @@ import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.Option;
 import net.geforcemods.securitycraft.api.Option.BooleanOption;
 import net.geforcemods.securitycraft.api.Option.DoubleOption;
+import net.geforcemods.securitycraft.api.Option.EnumOption;
 import net.geforcemods.securitycraft.api.Option.IntOption;
 import net.geforcemods.securitycraft.inventory.CustomizeBlockMenu;
 import net.geforcemods.securitycraft.items.ModuleItem;
@@ -244,6 +245,8 @@ public class CustomizeBlockScreen extends GuiContainer implements IContainerList
 	private String getValueText(Option<?> option) {
 		if (option instanceof BooleanOption)
 			return new TextComponentTranslation(((BooleanOption) option).get() ? "gui.securitycraft:invScan.yes" : "gui.securitycraft:invScan.no").getFormattedText();
+		else if (option instanceof EnumOption<?>)
+			return ((EnumOption<?>) option).getValueName().getFormattedText();
 		else
 			return option.toString();
 	}